TV: Prevent the live TV pop-up EPG from nesting and hanging the UI
If the 'start EPG' key (S) is pressed twice quickly right after starting
live TV then a Myth busy dialog box appears which can't be dismissed.

The problem is caused by attempting to start the pop-up EPG while it is
already active.

This patch checks if the pop-up is active and ignores the request if it is.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>