TV: Move ownership of the AudioOutputSync setting to PlayerContext
Currently the audio sync offset is setup in the MythPlayer ctor from
the AudioSyncOffset dbase setting.  However, this means that all
users of MythPlayer, including the transcoder, use the same offset.
This is OK for recordings and liveTV (where the user has interactively
set the value based on observed delays) but adversely affects transcoding
which results in permanent A/V offsets.

This patch moves ownership of AudioSyncOffset from MythPlayer
to PlayerContext.  PlayerContext sets the offset in StartPlaying and saves
it in StopPlaying.

Also the offset is set to 0 when using the cutlist editor to ensure
synchronisation between the audio output graph and the pause frame.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>