VO: Permit change of deinterlace method while playing
During playback there is a UI menu to select the deinterlace method.
However, VideoOutput::SetupDeinterlace does nothing if the interlacing
state is not changed.  This means the user is unable to change the
deinterlace filter during playback.

This patch adds  an additional test to VideoOutput::SetupDeinterlace
to allow a new deinterlace filter to be chosen.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>