TV: Ensure LiveTV channel number is updated when changing channel on same mux
When changing LiveTV channel from a DTV source, TVRec::TuningFrequency
checks if the new channel is on the same multiplex and if so it
skips the tuning operation and restarts the recorder.  Unfortunately
this means that the new ringbuffer filename will use the old channel
number.  This in turn means that the FE will use the wrong channel
number when disping OSD info.

This patch updates the channel number before restarting the recorder.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>