Player: Restore the cutlist editor after seeking across a video format change
A change in video format results in a call to MythPlayer::ReinitOSD
which reinitialises the OSD for the new video bounds.  However, if the
cutlist editor is active its window is left hidden until the user presses
a key.

This patch calls deleteMap.UpdateOSD from ReinitOSD to restore the window
and its contents.  NB deleteMap must be marked as changed for the
cut regions to be redrawn.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>