Mythwelcome: Disable the idle standby timer
Mythwelcome normally connects to the BE in non-blocking mode i.e. the BE
is free to shutdown when it's idle.

If mythwelcome is active and the MythMainWindow idle timer fires then
standby state is entered.  On the next key press, e.g. to start the FE,
standby mode is exited and gCoreContext->BlockShutdown is called which
prevents the BE from shutting down even when the FE enters standby.  The
block is not released until the FE exits and an idle timeout occurs in
mythwelcome.

Mythwelcome has no need for standby mode so disable the idle timer
during application startup.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>