mythsystemunix: Fix race condition in MythSystemLegacy::Run
After a child process is created in MythSystemLegacyUnix::Fork there
is a small time interval before MythSystemLegacy::Run calls
MythSystemLegacyUnix::Manage to add the process to
MythSystemLegacyManager's active queue.  If the child should terminate
before this then MythSystemLegacy::Wait can deadlock.

This fix acquires a mutex shared with MythSystemLegacyManager before
the fork operation and releases it after the process has been added
to the queue.

Additionally preserve the values of the pipe ends that are shared
with the child process.

The race condition and subsequent deadlock were observed in mythshutdown
with the --setwakeuptime option.  This prevented the BE shutting down
when idle and deadlocked the main BE thread.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>