BE: Avoid SEGV when accessing discarded MythSocket
1. When a client connects, MainServer::NewConnection is called which
inserts a new MythSocket in controlSocketList.

2. When data is received, MainServer::readyRead creates a new thread to
process the request and saves the MythSocket to read from.

3. If the client disconnects before the thread handles the request then
the MythSocket is added to decrRefSocketList and removed from
controlSocketList.

4. When the thread runs and handles an 'ANN' request it creates a
PlaybackSock that references the MythSocket.

5. When MainServer::customEvent runs, it deletes any MythSockets
listed in decrRefSocketList.  This results in the PlaybackSock having
a dangling reference to the socket.  If the event is to be broadcast,
all client PlaybackSock's are sent the message which results in a reference
to the deleted socket and a SEGV.

This fix adds a test in MainServer::ProcessRequestWork to ensure the
socket is still in the controlSocketList and some additional code
in MainServer::HandleAnnounce to also check controlSocketList before
a PlaybackSock is created.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>