UI: Ensure the UI idle timer is start/stopped from the UI thread
MythMainWindow::PauseIdleTimer is called from many places none of
which are guaranteed to be running on the main UI thread.  Qt
requires that timers are start/stopped from the same thread that
created them otherwise the operation will fail.

MusicPlayer::stop() calls MythMainWindow::PauseIdleTimer from
PlaylistLoadingThread which tickles this bug.

This patch causes the start/stop timer methods to be invoked
using QMetaObject which guarantees this requirement.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>